requirements.txt
streamlit==1.21.0
filterframes==0.1.1
biopython==1.81